changes int to long for relational object methods. Looks good and tests pass.

+++ plugins/org.komodo.relational/src/org/komodo/relational/RelationalConstants.java
@@ -77,11 +77,11 @@ public interface RelationalConstants {
     /**
      * The default value for the <code>datatype precision</code> property. Value is {@value} .
      */
-    int DEFAULT_PRECISION = 0;
+    long DEFAULT_PRECISION = 0;
 
     /**
      * The default value for the <code>datatype scale</code> property. Value is {@value} .
      */
-    int DEFAULT_SCALE = 0;
+    long DEFAULT_SCALE = 0;
 
 }
