+++ tests/org.komodo.relational.test/src/org/komodo/relational/model/internal/ResultSetColumnImplTest.java
@@ -225,20 +225,20 @@ public final class ResultSetColumnImplTest extends RelationalModelTest {
 
     @Test
     public void shouldSetDatatypePrecisionProperty() throws Exception {
-        final int value = 10;
+        final long value = 10;
         this.column.setPrecision( this.uow, value );
         assertThat( this.column.getPrecision( this.uow ), is( value ) );
         assertThat( this.column.getProperty( this.uow, StandardDdlLexicon.DATATYPE_PRECISION ).getLongValue( this.uow ),
-                    is( ( long )value ) );
+                    is( value ) );
     }
 
     @Test
     public void shouldSetDatatypeScaleProperty() throws Exception {
-        final int value = 10;
+        final long value = 10;
         this.column.setScale( this.uow, value );
         assertThat( this.column.getScale( this.uow ), is( value ) );
         assertThat( this.column.getProperty( this.uow, StandardDdlLexicon.DATATYPE_SCALE ).getLongValue( this.uow ),
-                    is( ( long )value ) );
+                    is( value ) );
     }
 
     @Test
